+++ packages/stream_chat/lib/src/client/retry_policy.dart
@@ -1,6 +1,6 @@
 import 'dart:async';
 
-import '../core/error/error.dart';
+import '../core/error/stream_chat_exception.dart';
 import 'client.dart';
 
 /// The retry policy associated to a client.