+++ packages/stream_chat/lib/src/core/error/error.dart
@@ -1,2 +1,0 @@
-export 'stream_chat_error.dart';
-export 'stream_chat_exception.dart';